+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ﻿# Latest Changelog
 
+- Scripts: Implemented 'setInterval' and 'setTimeout'
+- Scripts: Added audio resampling for the Audio API, audio files with 48khz should work now.
+
 - NEW!: Added JavaScript plugins support, you can now extend this tool by creating your own scripts that can also be shared.
   * JavaScript documentation can be found in the Github repo, under Docs/Scripts.md
   * This feature is still on an experimental phase and requires advanced JS understanding. Join the discord linked in the README for feedback, suggestions or requests.